Refactor: Move acp/ into core/src/acp/ module
Merged the standalone `acp` crate into `core` as a submodule to reduce
root-level directories and improve code organization.

Changes:
- Moved acp/src/* to core/src/acp/
- Updated core/src/lib.rs to export acp module
- Updated ts-rs export paths from ../../client/ to ../../../client/
- Removed acp from workspace members in root Cargo.toml
- Removed reprod-acp workspace dependency
- Added acp-specific dependencies to core/Cargo.toml:
  - agent-client-protocol
  - tokio-util
  - which
  - zip
- Updated desktop/Cargo.toml and server/Cargo.toml to remove reprod-acp
- Updated all imports from reprod_acp:: to reprod_core::acp::
- Fixed internal acp module imports to use super:: and crate::
- Deleted standalone acp/ directory

ACP functionality is now accessed via `reprod_core::acp` instead of
`reprod_acp`, simplifying the dependency graph while maintaining the
same functionality.

Verified with: cargo check ✓
