--- acp/Cargo.toml
@@ -1,25 +1,0 @@
-[package]
-name = "reprod-acp"
-version.workspace = true
-edition.workspace = true
-
-[dependencies]
-reprod-core = { workspace = true }
-anyhow = { workspace = true }
-async-trait = { workspace = true }
-agent-client-protocol = "0.9.0"
-dunce = "1.0"
-flate2 = "1.0"
-reqwest = { workspace = true }
-serde = { workspace = true }
-serde_json = { workspace = true }
-tar = "0.4"
-ts-rs = { version = "9.0", features = ["serde-json-impl"] }
-tokio = { workspace = true }
-tokio-util = { version = "0.7", features = ["compat"] }
-tracing = { workspace = true }
-which = "6.0"
-zip = { version = "2.2", default-features = false, features = ["deflate"] }
-
-[lints]
-workspace = true
