+++ core/Cargo.toml
@@ -33,6 +33,11 @@ dunce = "1.0"
 image = { version = "0.24", default-features = false, features = ["png", "jpeg", "gif", "bmp"] }
 printpdf = { version = "0.7", features = ["embedded_images"] }
 ts-rs = { version = "9.0", features = ["serde-json-impl", "chrono-impl"] }
+# ACP module dependencies
+agent-client-protocol = "0.9.0"
+tokio-util = { version = "0.7", features = ["compat"] }
+which = "6.0"
+zip = { version = "2.2", default-features = false, features = ["deflate"] }
 
 [target.'cfg(unix)'.dependencies]
 nix = { version = "0.27", features = ["signal"] }
