+++ core/src/acp/config.rs
@@ -1,7 +1,7 @@
 use std::path::PathBuf;
 
 use anyhow::{bail, Result};
-use reprod_core::config::{app_config_dir, workspace_root_override};
+use crate::config::{app_config_dir, workspace_root_override};
 use serde::{Deserialize, Serialize};
 use tracing::{error, info, warn};
 
@@ -152,7 +152,7 @@ pub fn is_external_mode(mode: &str) -> bool {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use reprod_core::config::{APP_DIR_ENV, WORKSPACE_ROOT_ENV};
+    use crate::config::{APP_DIR_ENV, WORKSPACE_ROOT_ENV};
     use std::{env, fs, sync::Mutex};
 
     static ENV_LOCK: Mutex<()> = Mutex::new(());
