+++ server/Cargo.toml
@@ -5,7 +5,6 @@ edition.workspace = true
 
 [dependencies]
 reprod-core = { workspace = true }
-reprod-acp = { workspace = true }
 
 # Web framework
 axum = { workspace = true }
